# shared/openapi.yaml
openapi: 3.1.0
info:
  title: ThreadLoop API
  version: 0.1.0
  description: |
    Source-of-truth contract for the ThreadLoop marketplace API.
    The TypeScript types in `shared/src/types` mirror this spec.
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://api.threadloop.example
    description: Production
...
paths:
...
  /api/auth/{provider}/callback:
    post:
      tags: [auth]
      summary: Exchange a provider credential for a ThreadLoop session
      description: |
        Single endpoint that dispatches by `provider` path parameter.
        The request body shape varies by provider (see `SsoCallbackInput`):
          - `google`:   `{ id_token }`
          - `apple`:    `{ id_token, code, name? }`
          - `facebook`: `{ access_token }`

        On success, returns a `Session` and sets a `Set-Cookie: refresh_token=…`
        header (httpOnly, Secure, SameSite=Lax, 30-day rolling lifetime).

        On a verified-email collision with an existing account from a different
        provider, returns `200` with `link_required: true` and a short-lived
        `link_token`; the client should prompt the user to re-authenticate
        with the original provider to confirm linking.
      parameters:
        - in: path
          name: provider
          required: true
          schema: { $ref: "#/components/schemas/AuthProvider" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SsoCallbackInput" }
      responses:
        "200":
          description: |
            Session created (or pending-link state — see `link_required`).
            Refresh token set as httpOnly cookie when not pending-link.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: Malformed request (missing required field for provider).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Provider token signature invalid or expired.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Unknown provider in path.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Provider JWKS unreachable; client should retry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, description: "Stable machine-readable error code." }
        message: { type: string, description: "Human-readable detail; safe to surface." }
        request_id:
          type: string
          description: "Server-assigned request id for log correlation."
    AuthProvider:
      type: string
      enum: [google, apple, facebook]
    GoogleSsoCallbackInput:
      type: object
      required: [id_token]
      properties:
        id_token: { type: string, description: "Google-issued ID token (JWT)." }
    AppleSsoCallbackInput:
      type: object
      required: [id_token, code]
      properties:
        id_token: { type: string, description: "Apple-issued ID token (JWT)." }
        code:     { type: string, description: "Apple authorization code; used to redeem an Apple refresh token server-side." }
        name:
          type: string
          nullable: true
          description: |
            User's display name as surfaced by Apple's JS SDK / native flow on
            the first sign-in only (and only when the app requested the `name`
            scope). Subsequent sign-ins omit it entirely; the backend reuses
            the existing `users.display_name` in that case.
    FacebookSsoCallbackInput:
      type: object
      required: [access_token]
      properties:
        access_token: { type: string, description: "Facebook user access token (resolved against Graph API server-side)." }
    SsoCallbackInput:
      description: |
        Provider-specific request body. Backend validates against the variant
        matching the `provider` path parameter.
      oneOf:
        - $ref: "#/components/schemas/GoogleSsoCallbackInput"
        - $ref: "#/components/schemas/AppleSsoCallbackInput"
        - $ref: "#/components/schemas/FacebookSsoCallbackInput"
    User:
      type: object
      required: [id, provider, display_name, email_verified, can_sell, can_purchase, created_at, updated_at]
      properties:
        id:             { type: string, format: uuid }
        provider:       { $ref: "#/components/schemas/AuthProvider" }
        email:          { type: string, nullable: true, description: "May be null when Apple withholds the address." }
        email_verified: { type: boolean }
        display_name:   { type: string }
        avatar_url:     { type: string, nullable: true, format: uri }
        can_sell:       { type: boolean }
        can_purchase:   { type: boolean }
        seller_rating:  { type: number, nullable: true, minimum: 0, maximum: 5 }
        created_at:     { type: string, format: date-time }
        updated_at:     { type: string, format: date-time }
    Session:
      type: object
      description: |
        Result of a successful authentication. When `link_required` is `true`,
        the existing-account collision flow has fired: `access_token`/`user`
        are absent and the client must surface a "sign in with {existing
        provider} to link" prompt, then submit `link_token` alongside the
        re-auth result. On the happy path, `link_required` is `false` (or
        omitted) and `access_token`/`expires_at`/`user` are all present.
      required: [link_required]
      properties:
        access_token:
          type: string
          description: "Short-lived (15 min) access JWT. Absent when link_required is true."
        expires_at:
          type: string
          format: date-time
          description: "Access token expiry. Absent when link_required is true."
        user:
          allOf:
            - $ref: "#/components/schemas/User"
          description: "Authenticated user. Absent when link_required is true."
        link_required:
          type: boolean
          default: false
          description: "True when the email collides with an existing account from a different provider; client must prompt the user to re-authenticate with that provider to link."
        link_provider:
          $ref: "#/components/schemas/AuthProvider"
          description: "Existing provider the user must re-authenticate with. Present iff link_required is true."
        link_token:
          type: string
          description: "Short-lived token submitted alongside the linking re-auth to confirm the merge. Present iff link_required is true."
...